Added missing relation for users and user deletion (self and by an admin)

+++ 03_backend_node_mongo_api_rest/p7/src/api/controllers/users.js
@@ -1,5 +1,6 @@
 const { generateSign } = require("../../config/jwt");
 const User = require("../models/users");
+const Libro = require("../models/libros");
 const bcrypt = require("bcrypt");
 
 const buscarUsuario = async (userName) => {
@@ -9,9 +10,11 @@ const buscarUsuario = async (userName) => {
 
 const getUsers = async (req, res, next) => {
   try {
-    const users = await User.find();
+    const users = await User.find().populate("favoritos");
     return res.status(200).json(users);
   } catch (error) {
+    let err_msg = "Error en la solicitud 'getUsers'. Error: " + error;
+    console.error(err_msg);
     return res.status(400).json(error);
   }
 }
@@ -35,6 +38,8 @@ const register = async (req, res, next) => {
     return res.status(201).json(userSaved);
 
   } catch (error) {
+    let err_msg = "Error en la solicitud 'register'. Error: " + error;
+    console.error(err_msg);
     return res.status(400).json(error);
   }
 }
@@ -54,8 +59,9 @@ const login = async (req, res, next) => {
     } else {
       return res.status(400).json("Contraseña incorrecta");
     }
-
   } catch (error) {
+    let err_msg = "Error en la solicitud 'login'. Error: " + error;
+    console.error(err_msg);
     return res.status(400).json(error);
   }
 }
@@ -77,8 +83,19 @@ const putUser = async (req, res, next) => {
 const deleteUser = async (req, res, next) => {
   try {
     const { id } = req.params;
-    const userDeleted = await User.findByIdAndDelete(id);
-    return res.status(200).json(userDeleted);
+    const requestingUser = req.userName;
+
+    if (requestingUser.rol === 'admin' || requestingUser._id === id) {
+      const userDeleted = await User.findByIdAndDelete(id);
+
+      if (userDeleted) {
+        return res.status(200).json({ message: "Usuario eliminado correctamente", userDeleted });
+      } else {
+        return res.status(404).json({ message: "Usuario no encontrado" });
+      }
+    } else {
+      return res.status(403).json({ message: "No tienes permisos para realizar esta acción" });
+    }
   } catch (error) {
     let err_msg = "Error en la solicitud 'deleteUser'. Error: " + error;
     console.error(err_msg);
@@ -86,4 +103,54 @@ const deleteUser = async (req, res, next) => {
   }
 }
 
-module.exports = { getUsers, register, login, putUser, deleteUser }
+const getFavoritos = async (req, res) => {
+  try {
+    const user = await User.findById(req.user._id).populate('favoritos');
+    return res.status(200).json(user.favoritos);
+  } catch (error) {
+    let err_msg = "Error en la solicitud 'getFavoritos'. Error: " + error;
+    console.error(err_msg);
+    return res.status(400).json(err_msg);
+  }
+}
+
+const addFavorito = async (req, res) => {
+  try {
+    const { id } = req.body;
+    const libro = await Libro.findById(id);
+
+    if (!libro) {
+      return res.status(404).json("Libro no encontrado");
+    }
+
+    const user = await User.findById(req.user._id);
+    if (!user.favoritos.includes(id)) {
+      user.favoritos.push(id);
+      await user.save();
+    }
+
+    return res.status(200).json(user.favoritos);
+  } catch (error) {
+    let err_msg = "Error en la solicitud 'addFavorito'. Error: " + error;
+    console.error(err_msg);
+    return res.status(400).json(err_msg);
+  }
+}
+
+const deleteFavorito = async (req, res) => {
+  try {
+    const { id } = req.params;
+    const user = await User.findById(req.user._id);
+
+    user.favoritos.pull(id);
+    await user.save();
+
+    return res.status(200).json(user.favoritos);
+  } catch (error) {
+    let err_msg = "Error en la solicitud 'deleteFavorito'. Error: " + error;
+    console.error(err_msg);
+    return res.status(400).json(err_msg);
+  }
+}
+
+module.exports = { getUsers, register, login, putUser, deleteUser, getFavoritos, addFavorito, deleteFavorito };
